highlights: add highlights immediately
